requirements.txt
asgiref==3.2.3
Django==3.0.2
django-admin-honeypot==1.1.0
django-blacklist==0.3.0
django-bot-crawler-blocker==1.0
django-ratelimit==2.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
